+++ data_labeling_examples/bulk_labeling_java/src/main/java/com/oracle/datalabelingservicesamples/tasks/CreateAnnotationTask.java
@@ -2,7 +2,7 @@ package com.oracle.datalabelingservicesamples.tasks;
 
 import com.oracle.bmc.datalabelingservicedataplane.model.Annotation;
 import com.oracle.bmc.datalabelingservicedataplane.model.CreateAnnotationDetails;
-import com.oracle.datalabelingservicesamples.utils.DataPlaneAPIWrapper;
+import com.oracle.datalabelingservicesamples.utils.DlsApiWrapper;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.Optional;
@@ -10,23 +10,23 @@ import java.util.Optional;
 @Slf4j
 public class CreateAnnotationTask extends Tasks<Annotation> {
     final String opcRequestId;
-    final DataPlaneAPIWrapper dataPlaneAPIWrapper;
+    final DlsApiWrapper dlsApiWrapper;
     private final CreateAnnotationDetails createAnnotationDetails;
 
     public CreateAnnotationTask(
             CreateAnnotationDetails createAnnotationDetails,
             String opcRequestId,
-            DataPlaneAPIWrapper dataPlaneAPIWrapper) {
+            DlsApiWrapper dlsApiWrapper) {
         this.opcRequestId = opcRequestId;
         this.createAnnotationDetails = createAnnotationDetails;
-        this.dataPlaneAPIWrapper = dataPlaneAPIWrapper;
+        this.dlsApiWrapper = dlsApiWrapper;
     }
 
     @Override
     public Annotation call() throws Exception {
         log.info("Creating Annotation for record {}", createAnnotationDetails.getRecordId());
         try {
-            return dataPlaneAPIWrapper.createAnnotation(
+            return dlsApiWrapper.createAnnotation(
                     createAnnotationDetails,
                     Optional.of(opcRequestId));
         } catch (Exception e) {
