+++ data_labeling_examples/bulk_labeling_java/src/main/java/com/oracle/datalabelingservicesamples/tasks/TaskHandler.java
@@ -5,7 +5,7 @@ import com.oracle.bmc.datalabelingservicedataplane.model.CreateAnnotationDetails
 import com.oracle.bmc.datalabelingservicedataplane.model.RecordSummary;
 import com.oracle.datalabelingservicesamples.labelingstrategies.MlAssistedLabelingStrategy;
 import com.oracle.datalabelingservicesamples.requests.AssistedLabelingParams;
-import com.oracle.datalabelingservicesamples.utils.DataPlaneAPIWrapper;
+import com.oracle.datalabelingservicesamples.utils.DlsApiWrapper;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
@@ -95,7 +95,7 @@ public class TaskHandler {
 
     public List<Future<Annotation>> getCreateAnnotationTasks(
             List<CreateAnnotationDetails> createAnnotationDetailsList,
-            DataPlaneAPIWrapper dataPlaneAPIWrapper,
+            DlsApiWrapper dlsApiWrapper,
             String opcRequestId,
             ExecutorService executorService) {
         return createAnnotationDetailsList.stream()
@@ -105,7 +105,7 @@ public class TaskHandler {
                                         mdcWrap(
                                                 taskProvider.getCreateAnnotationTask(
                                                         createAnnotationDetails,
-                                                        dataPlaneAPIWrapper,
+                                                        dlsApiWrapper,
                                                         opcRequestId))))
                 .collect(Collectors.toList());
     }
