+++ data_labeling_examples/bulk_labeling_java/src/main/java/com/oracle/datalabelingservicesamples/tasks/TaskProvider.java
@@ -4,17 +4,17 @@ import com.oracle.bmc.datalabelingservicedataplane.model.CreateAnnotationDetails
 import com.oracle.bmc.datalabelingservicedataplane.model.RecordSummary;
 import com.oracle.datalabelingservicesamples.labelingstrategies.MlAssistedLabelingStrategy;
 import com.oracle.datalabelingservicesamples.requests.AssistedLabelingParams;
-import com.oracle.datalabelingservicesamples.utils.DataPlaneAPIWrapper;
+import com.oracle.datalabelingservicesamples.utils.DlsApiWrapper;
 
 import java.util.List;
 
 public class TaskProvider {
     public CreateAnnotationTask getCreateAnnotationTask(
             CreateAnnotationDetails createAnnotationDetails,
-            DataPlaneAPIWrapper dataPlaneAPIWrapper,
+            DlsApiWrapper dlsApiWrapper,
             String opcRequestId) {
         return new CreateAnnotationTask(
-                createAnnotationDetails, opcRequestId, dataPlaneAPIWrapper);
+                createAnnotationDetails, opcRequestId, dlsApiWrapper);
     }
 
     public CreateBulkAnalyzeTask provideAssistedLabellingTask(
