+++ data_labeling_examples/bulk_labeling_java/src/main/resources/config.properties
@@ -15,7 +15,7 @@ REGION=us-phoenix-1
 #DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iacky2zth6bor2gb2mzyr2yhnszexprfehs5xuq4fand6a
 
 #Text classification multi label
-#DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iazmjbyddco6dxp5yotpj3gts5pheapdyafzr2kjramaaq
+DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iazmjbyddco6dxp5yotpj3gts5pheapdyafzr2kjramaaq
 
 #Image classification single label
 #DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iafag3p4tgg42fm4u6yfmpwmmo2udstg6a53dv4b3gqoqq
@@ -24,7 +24,7 @@ REGION=us-phoenix-1
 #DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iarcinopn7s4dm4g5t5hldqwe5h2g34mao7swvfowqxpta
 
 #Obj detection custom model
-DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iaxy5xn2y24tetjy6d5rtlplbzz6nxzsmsym5z4u5wdvpa
+#DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iaxy5xn2y24tetjy6d5rtlplbzz6nxzsmsym5z4u5wdvpa
 
 #Entity extraction custom model
 #DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iant6qouko5xwlhtcb467yxx2toomn2sgug4kigwmjxmcq
@@ -56,6 +56,9 @@ CUSTOM_MODEL_ID=
 
 #MODEL_TRAINING_PROJECT_ID=ocid1.aivisionproject.oc1.phx.amaaaaaaniob46iaf4m7a2kgs7ysguaycarzmewo4ksu45wwwapv6n5h47kq
 
-MODEL_TRAINING_PROJECT_ID=
+MODEL_TRAINING_PROJECT_ID=ocid1.ailanguageproject.oc1.phx.amaaaaaaniob46iaci6amgp3lnvpbr5iczdlvfymnfdpecflpkor4cipe5iq
 
-TRAINING_DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iacwjv4kjtswmi2oar4vikzdqx4f7wtxdbrwhw3vnx6vqq
+#TRAINING_DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iacwjv4kjtswmi2oar4vikzdqx4f7wtxdbrwhw3vnx6vqq
+
+#Text classification
+TRAINING_DATASET_ID=ocid1.datalabelingdatasetint.oc1.phx.amaaaaaaniob46iacky2zth6bor2gb2mzyr2yhnszexprfehs5xuq4fand6a
